+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Include source file path in GitHub Action PR body table
+
 ## [1.1.4] - 2026-06-19
 
 ### Fixed